requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
pydantic-settings 
python-multipart